lowerNonArrayAggregate affected by comment in cond.d so moved to statementsem
always remeber that when you move a function, the function still needs to be used and the purpose still has to be seen in the whole dmd unit. but since it has been moved to a new location, how it is accessed must change. it needs to be access from its new location whiles complying to the semantic rule of programming.