move prepare to statementsem.
before you look at this commit, go back and look into prepare function.
it calls lowerNonArrayAggregate frunction. but since they are in the same class, it can directly just call it
without any special means. just call it and use it.

but since we've moved lowerNonArrayAggregate, it will break because it can no longer find it in the class. and since it also depends on semantic routines, also needs to be moved to statementsem so it can nicely get to call lowerNonArrayAggregate