do local import and change how prepare is called in attrib.d
remember that before moving prepare, it was a member of a class. so must be called from an instance of that class.
that's the only way you can call a member function.
but prepare is no longer in the class. so we access it from where it is and we call it.
it is now in statementsem so we need to do a local import of it and call it as a standalone function as that is what it is now